+++ mall-order-service/src/main/java/cn/enjoy/mall/dao/DeliveryDocMapper.java
@@ -61,5 +61,5 @@ public interface DeliveryDocMapper {
      */
     int updateByPrimaryKey(DeliveryDoc record);
 
-    List<DeliveryDoc> selectByOrderId(int orderId);
+    List<DeliveryDoc> selectByOrderId(Long orderId);
 }
